refactor: correct naming
