+++ backend/src/routing/accountRouter.ts
@@ -1,11 +1,12 @@
 import { Router } from 'express';
 
 import { SubRoutes } from 'enums/routesEnum';
-import { loginController, refreshController } from 'controllers/auth/authController';
+import login from 'controllers/auth/login';
+import refresh from 'controllers/auth/refresh';
 
 const accountRouter = Router();
 
-accountRouter.post(`${SubRoutes.login}`, loginController);
-accountRouter.get(`${SubRoutes.refresh}`, refreshController);
+accountRouter.post(`${SubRoutes.login}`, login);
+accountRouter.get(`${SubRoutes.refresh}`, refresh);
 
 export default accountRouter;
